remove all graphql references from components
