+++ packages/component-submission/client/components/SubmissionOperations.js
@@ -1,10 +1,7 @@
 import React from 'react'
-import { graphql, compose } from 'react-apollo'
-import { ALL_MANUSCRIPTS } from '@elifesciences/component-dashboard/client/graphql/queries'
 import { Loading } from '@elifesciences/component-elife-ui/client/atoms'
 import { ErrorPage } from '@elifesciences/component-elife-app/client'
 
-import { UPDATE_MANUSCRIPT, SUBMIT_MANUSCRIPT } from './operations'
 import { parseFormToOutputData, parseInputToFormData } from '../utils'
 
 const SubmissionOperations = ({
@@ -26,18 +23,13 @@ const SubmissionOperations = ({
     initialValues: parseInputToFormData(data.manuscript),
     updateManuscript: formValues =>
       updateManuscript({
-        refetchQueries: [{ query: ALL_MANUSCRIPTS }],
         variables: { data: parseFormToOutputData(formValues) },
       }),
     submitManuscript: formValues =>
       submitManuscript({
-        refetchQueries: [{ query: ALL_MANUSCRIPTS }],
         variables: { data: parseFormToOutputData(formValues) },
       }),
   })
 }
 
-export default compose(
-  graphql(UPDATE_MANUSCRIPT, { name: 'updateManuscript' }),
-  graphql(SUBMIT_MANUSCRIPT, { name: 'submitManuscript' }),
-)(SubmissionOperations)
+export default SubmissionOperations
